Fixed a bug in totalCost(), changed type to Bigdecimal; more appropriate for financial calculations.

+++ src/com/main/workerclasses/DiscountComputer.java
@@ -72,7 +72,7 @@ public class DiscountComputer {
            totalSavings = totalDiscountSavings.add(totalSavings);
        }
 
-       BigDecimal basketTotal = BigDecimal.valueOf(passedBasket.totalCost());
+       BigDecimal basketTotal = passedBasket.totalCost();
 
        BigDecimal postTotalDiscount = basketTotal.subtract(totalSavings);
